Refactored to make some code more generic
