+++ MeNext.Spotify/SpotifyMetadata/JsonResults/IResultWithList.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace MeNext.Spotify
+{
+    public interface IResultWithList<T>
+    {
+        IList<T> Items { get; }
+    }
+}
