+++ MeNext.Spotify/SpotifyMetadata/JsonResults/TracksResult.cs
@@ -3,8 +3,16 @@ using System.Collections.Generic;
 
 namespace MeNext.Spotify
 {
-    public class TracksResult
+    public class TracksResult : IResultWithList<TrackResult>
     {
+        public IList<TrackResult> Items
+        {
+            get
+            {
+                return tracks;
+            }
+        }
+
         public IList<TrackResult> tracks { get; set; }
     }
